Change name of the modules key
We were using the `:jlink-module-path` for the list of additional Java modules
and this clashes with Leiningen's handling of project keys. Lein assumes that
any project key containing `paths` is a list of paths on the file system and
any key containing `path` is a string with a path on the file system. In this
case our key includes `path` but contains a list causing Leiningen to throw an
exception.

https://github.com/technomancy/leiningen/blob/master/leiningen-core/src/leiningen/core/project.clj#L472